Appコンポーネント,Formコンポーネントの修正
