+++ src/components/Form.tsx
@@ -1,17 +1,13 @@
-import { useState } from 'react';
+type FormPropsType = {
+    setCity:React.Dispatch<React.SetStateAction<string>>;
+    getWeather: (e: any) => void;
+}
 
-const Form = () => {
-    const [city,setCity] = useState<string>("");
-    const getWeather = (e:any) => { 
-        e.preventDefault();
-        fetch("http://api.weatherapi.com/v1/current.json?key=0dfb86719d8349e797b142110211008&q=London&aqi=no")
-             .then( res => res.json())
-             .then(data =>console.log(data))
-    }    
+const Form = (props: FormPropsType) => {
     return(
         <form>
-            <input type="text" name="city" placeholder="都市名" onChange={e => setCity(e.target.value)}/>
-            <button type="submit" onClick={getWeather}>Get Weather</button>
+            <input type="text" name="city" placeholder="都市名" onChange={e => props.setCity(e.target.value)}/>
+            <button type="submit" onClick={props.getWeather}>Get Weather</button>
         </form>
     );
 };
